# [JSC] Optimize runtime iterations for arguments and string